+++ app/src/main/kotlin/com/d4rk/cartcalculator/ui/components/animations/Animations.kt
@@ -2,8 +2,11 @@ package com.d4rk.cartcalculator.ui.components.animations
 
 import android.content.Context
 import androidx.compose.animation.core.animateFloatAsState
+import androidx.compose.animation.core.tween
 import androidx.compose.foundation.gestures.awaitFirstDown
 import androidx.compose.foundation.gestures.waitForUpOrCancellation
+import androidx.compose.foundation.layout.offset
+import androidx.compose.foundation.layout.padding
 import androidx.compose.material3.DrawerState
 import androidx.compose.material3.SwipeToDismissBoxState
 import androidx.compose.material3.SwipeToDismissBoxValue
@@ -22,6 +25,8 @@ import androidx.compose.ui.hapticfeedback.HapticFeedbackType
 import androidx.compose.ui.input.pointer.pointerInput
 import androidx.compose.ui.platform.LocalContext
 import androidx.compose.ui.platform.LocalHapticFeedback
+import androidx.compose.ui.unit.IntOffset
+import androidx.compose.ui.unit.dp
 import com.d4rk.cartcalculator.data.datastore.DataStore
 import com.d4rk.cartcalculator.data.model.ui.button.ButtonState
 
@@ -80,18 +85,47 @@ fun Modifier.hapticDrawerSwipe(drawerState : DrawerState) : Modifier = composed 
     return@composed this
 }
 
-fun Modifier.hapticSwipeToDismissBox(swipeToDismissBoxState: SwipeToDismissBoxState): Modifier = composed {
-    val haptic: HapticFeedback = LocalHapticFeedback.current
-    var hasVibrated by remember { mutableStateOf(value = false) }
+fun Modifier.hapticSwipeToDismissBox(swipeToDismissBoxState : SwipeToDismissBoxState) : Modifier =
+        composed {
+            val haptic : HapticFeedback = LocalHapticFeedback.current
+            var hasVibrated by remember { mutableStateOf(value = false) }
 
-    LaunchedEffect(swipeToDismissBoxState.currentValue) {
-        if (swipeToDismissBoxState.currentValue != SwipeToDismissBoxValue.Settled && !hasVibrated) {
-            haptic.performHapticFeedback(HapticFeedbackType.LongPress)
-            hasVibrated = true
-        } else if (swipeToDismissBoxState.currentValue == SwipeToDismissBoxValue.Settled) {
-            hasVibrated = false
+            LaunchedEffect(swipeToDismissBoxState.currentValue) {
+                if (swipeToDismissBoxState.currentValue != SwipeToDismissBoxValue.Settled && ! hasVibrated) {
+                    haptic.performHapticFeedback(HapticFeedbackType.LongPress)
+                    hasVibrated = true
+                }
+                else if (swipeToDismissBoxState.currentValue == SwipeToDismissBoxValue.Settled) {
+                    hasVibrated = false
+                }
+            }
+
+            return@composed this
         }
-    }
 
-    return@composed this
+fun Modifier.animateVisibility(
+    visible : Boolean = true ,
+    index : Int = 0 ,
+    offsetY : Int = 50 ,
+    durationMillis : Int = 300 ,
+    delayPerItem : Int = 64
+) = composed {
+    val alpha = animateFloatAsState(
+        targetValue = if (visible) 1f else 0f , animationSpec = tween(
+            durationMillis = durationMillis , delayMillis = index * delayPerItem
+        ) , label = "Alpha"
+    )
+
+    val offsetYState = animateFloatAsState(
+        targetValue = if (visible) 0f else offsetY.toFloat() , animationSpec = tween(
+            durationMillis = durationMillis , delayMillis = index * delayPerItem
+        ) , label = "OffsetY"
+    )
+
+    this
+            .offset { IntOffset(x = 0 , offsetYState.value.toInt()) }
+            .graphicsLayer {
+                this.alpha = alpha.value
+            }
+            .padding(vertical = 4.dp)
 }